% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{dynamo_internal}
\alias{dynamo_internal}
\alias{design}
\title{Function that constructs design components.}
\usage{
...
}
\description{
Internal design function.
}
\details{
This function is not intended for use by users.
}
\author{
Adam Lund

################# construct design components
}
\keyword{internal}
