\name{motifclstest}
\alias{motifclstest}
\title{
Find motifs enriched in clustering
}
\description{
Calculate enrichment levels of motifs in clusters generated by K-means
clustering based on read counts at predicted motif sites across different
samples.
}
\usage{
motifclstest(motifcls, motifname)
}
\arguments{
  \item{motifcls}{
A matrix object with each row representing a motif site and each column
representing read counts from one sample. The last three columns are the ids
of clusters, the ids of motifs and the ids of motif sites.
}
  \item{motifname}{
A vector of ids of motifs. For example, if x motifs are examined, motifname is
1:x
}
}
\details{
This function counts the numbers of motif sites in each cluster for each
motif and calculates the fold enrichments and p values for each motif in each
cluster and outputs in a matrix where each row represents a motif.
}
\value{
A table is written in the "motif_cluster_enrichment.txt".
}
\author{
Zheng Kuang
}
\examples{
motiffdrread=read.table(system.file("extdata","motif_filter_reads_cluster.txt",
package="DynaMO",mustWork=TRUE),header=TRUE,sep="\t")
motiffdrclstest=motifclstest(motiffdrread,1:5)
}
\keyword{ motifclstest }
