% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V.R
\docType{data}
\name{V}
\alias{V}
\title{VSD imaging of a live ferret brain.}
\format{A three dimensional \eqn{25\times 25 \times 977} array.}
\description{
This data set contains one trial of processed voltage sentisitve dye reocordings from animal 308. 
The original raw data set contains integer values indicitating the flourescence. See \cite{Roland et. al, 2006}
}
\references{
Per E. Roland, Akitoshi Hanazawa, Calle Undeman, David Eriksson, Tamas Tompa, Hiroyuki Nakamura, Sonata Valentiniene and Bashir Ahmed (2006). Cortical feedback depolarization waves: A mechanism of top-down influence on early visual areas.
 \emph{PNAS}, 103, 33, 12586-12591.  url = {https://doi.org/10.1073/pnas.0604925103}.
}
